Changed actuation model structure. Instead of different base classes there are now "ActuationModel" classes that handle those aspects of the environment
