+++ mimoEnv/envs/muscle_test.py
@@ -5,7 +5,7 @@ functions that returned fixed values. This allows for testing the model without 
 The second class :class:`~mimoEnv.envs.dummy.MIMoShowroomEnv` is identical to the first, but changes the default
 parameters to load the showroom scene instead.
 
-Finally there is a demo class for the v2 version of MIMo using five-fingered hands and feet with two toes each in
+Finally, there is a demo class for the v2 version of MIMo using five-fingered hands and feet with two toes each in
 :class:`~mimoEnv.envs.dummy.MIMoV2DemoEnv`.
 """
 
@@ -14,9 +14,10 @@ import os
 import mujoco_py
 
 from mimoEnv.envs.mimo_env import SCENE_DIRECTORY, DEFAULT_VESTIBULAR_PARAMS, DEFAULT_PROPRIOCEPTION_PARAMS, MIMoEnv
-from mimoEnv.envs.mimo_muscle_env import MIMoMuscleEnv
 from mimoTouch.touch import TrimeshTouch
 import mimoEnv.utils as env_utils
+from mimoActuation.actuation import ActuationModel, TorqueMotorModel
+from mimoActuation.muscle import MuscleModel
 
 STATIC_TEST_XML = os.path.join(SCENE_DIRECTORY, "muscle_static_test.xml")
 """ Path to the benchmarking scene using MIMo v2.
@@ -46,17 +47,11 @@ COMPLIANCE_XML = os.path.join(SCENE_DIRECTORY, "compliance_test_scene.xml")
 COMPLIANCE_INIT_POSITION = {"robot:right_shoulder_ad_ab": np.array([1.35]), }
 
 
-class MIMoStaticMuscleTestEnv(MIMoMuscleEnv):
+class MIMoStaticMuscleTestEnv(MIMoEnv):
     def __init__(self,
                  model_path=STATIC_TEST_XML,
                  initial_qpos={},
                  n_substeps=2,
-                 proprio_params=DEFAULT_PROPRIOCEPTION_PARAMS,
-                 touch_params=None,
-                 vision_params=None,
-                 vestibular_params=DEFAULT_VESTIBULAR_PARAMS,
-                 goals_in_observation=False,
-                 done_active=True,
                  show_sensors=False,
                  print_space_sizes=False,):
 
@@ -66,12 +61,13 @@ class MIMoStaticMuscleTestEnv(MIMoMuscleEnv):
         super().__init__(model_path=model_path,
                          initial_qpos=initial_qpos,
                          n_substeps=n_substeps,
-                         proprio_params=proprio_params,
-                         touch_params=touch_params,
-                         vision_params=vision_params,
-                         vestibular_params=vestibular_params,
-                         goals_in_observation=goals_in_observation,
-                         done_active=done_active)
+                         proprio_params=None,
+                         touch_params=None,
+                         vision_params=None,
+                         vestibular_params=None,
+                         actuation_model=MuscleModel,
+                         goals_in_observation=False,
+                         done_active=False)
 
         if print_space_sizes:
             print("Observation space:")
@@ -173,24 +169,12 @@ class MIMoStaticMuscleTestV2Env(MIMoStaticMuscleTestEnv):
                  model_path=STATIC_TEST_XML_V2,
                  initial_qpos={},
                  n_substeps=2,
-                 proprio_params=DEFAULT_PROPRIOCEPTION_PARAMS,
-                 touch_params=None,
-                 vision_params=None,
-                 vestibular_params=DEFAULT_VESTIBULAR_PARAMS,
-                 goals_in_observation=False,
-                 done_active=True,
                  show_sensors=False,
                  print_space_sizes=False,):
 
         super().__init__(model_path=model_path,
                          initial_qpos=initial_qpos,
                          n_substeps=n_substeps,
-                         proprio_params=proprio_params,
-                         touch_params=touch_params,
-                         vision_params=vision_params,
-                         vestibular_params=vestibular_params,
-                         goals_in_observation=goals_in_observation,
-                         done_active=done_active,
                          show_sensors=show_sensors,
                          print_space_sizes=print_space_sizes,)
 
@@ -200,24 +184,12 @@ class MIMoVelocityMuscleTestEnv(MIMoStaticMuscleTestEnv):
                  model_path=VELOCITY_TEST_XML,
                  initial_qpos={},
                  n_substeps=2,
-                 proprio_params=DEFAULT_PROPRIOCEPTION_PARAMS,
-                 touch_params=None,
-                 vision_params=None,
-                 vestibular_params=DEFAULT_VESTIBULAR_PARAMS,
-                 goals_in_observation=False,
-                 done_active=True,
                  show_sensors=False,
                  print_space_sizes=False,):
 
         super().__init__(model_path=model_path,
                          initial_qpos=initial_qpos,
                          n_substeps=n_substeps,
-                         proprio_params=proprio_params,
-                         touch_params=touch_params,
-                         vision_params=vision_params,
-                         vestibular_params=vestibular_params,
-                         goals_in_observation=goals_in_observation,
-                         done_active=done_active,
                          show_sensors=show_sensors,
                          print_space_sizes=print_space_sizes,)
 
@@ -227,24 +199,12 @@ class MIMoVelocityMuscleTestV2Env(MIMoStaticMuscleTestEnv):
                  model_path=VELOCITY_TEST_XML_V2,
                  initial_qpos={},
                  n_substeps=2,
-                 proprio_params=DEFAULT_PROPRIOCEPTION_PARAMS,
-                 touch_params=None,
-                 vision_params=None,
-                 vestibular_params=DEFAULT_VESTIBULAR_PARAMS,
-                 goals_in_observation=False,
-                 done_active=True,
                  show_sensors=False,
                  print_space_sizes=False,):
 
         super().__init__(model_path=model_path,
                          initial_qpos=initial_qpos,
                          n_substeps=n_substeps,
-                         proprio_params=proprio_params,
-                         touch_params=touch_params,
-                         vision_params=vision_params,
-                         vestibular_params=vestibular_params,
-                         goals_in_observation=goals_in_observation,
-                         done_active=done_active,
                          show_sensors=show_sensors,
                          print_space_sizes=print_space_sizes,)
 
@@ -257,19 +217,17 @@ class MIMoComplianceEnv(MIMoEnv):
                  model_path=COMPLIANCE_XML,
                  initial_qpos=COMPLIANCE_INIT_POSITION,
                  n_substeps=2,
-                 proprio_params=DEFAULT_PROPRIOCEPTION_PARAMS,
-                 touch_params=None,
-                 vision_params=None,
-                 vestibular_params=None,
+                 actuation_model=TorqueMotorModel,
                  ):
 
         super().__init__(model_path=model_path,
                          initial_qpos=initial_qpos,
                          n_substeps=n_substeps,
-                         proprio_params=proprio_params,
-                         touch_params=touch_params,
-                         vision_params=vision_params,
-                         vestibular_params=vestibular_params,
+                         proprio_params=None,
+                         touch_params=None,
+                         vision_params=None,
+                         vestibular_params=None,
+                         actuation_model=actuation_model,
                          goals_in_observation=False,
                          done_active=False)
 
@@ -355,106 +313,16 @@ class MIMoComplianceEnv(MIMoEnv):
         self.viewer.cam.azimuth = 180
 
 
-class MIMoComplianceMuscleEnv(MIMoMuscleEnv):
+class MIMoComplianceMuscleEnv(MIMoComplianceEnv):
     """ Test environment for muscle adaptive compliance.
     """
     def __init__(self,
                  model_path=COMPLIANCE_XML,
                  initial_qpos=COMPLIANCE_INIT_POSITION,
                  n_substeps=2,
-                 proprio_params=DEFAULT_PROPRIOCEPTION_PARAMS,
-                 touch_params=None,
-                 vision_params=None,
-                 vestibular_params=None,
                  ):
 
         super().__init__(model_path=model_path,
                          initial_qpos=initial_qpos,
                          n_substeps=n_substeps,
-                         proprio_params=proprio_params,
-                         touch_params=touch_params,
-                         vision_params=vision_params,
-                         vestibular_params=vestibular_params,
-                         goals_in_observation=False,
-                         done_active=False)
-
-        joint_names = [self.sim.model.joint_id2name(joint_id) for joint_id in self.mimo_joints]
-        for joint_name in joint_names:
-            env_utils.lock_joint(self.sim.model, joint_name)
-        env_utils.unlock_joint(self.sim.model, "robot:right_shoulder_ad_ab")
-        env_utils.lock_joint(self.sim.model, "robot:right_hand1", joint_angle=-0.25)
-        # TODO: Determine two sets of muscle action inputs, one set of motor inputs.
-        #   Inputs should keep arm approximately horizontal
-        #   Then we drop ball on the arm and plot relevant units over time: qpos, qvel, torque
-        # Let sim settle for a few timesteps to allow weld and locks to settle
-        gravity = self.sim.model.opt.gravity[2]
-        self.sim.model.opt.gravity[2] = 0
-        self.do_simulation(np.zeros(self.action_space.shape), 2)
-        self.sim.model.opt.gravity[2] = gravity
-        self.init_qpos = self.sim.data.qpos.copy()
-
-    def _sample_goal(self):
-        """ Dummy function.
-        """
-        return np.zeros((0,))
-
-    def _is_success(self, achieved_goal, desired_goal):
-        """ Dummy function. Always returns False.
-        """
-        return False
-
-    def compute_reward(self, achieved_goal, desired_goal, info):
-        """ Dummy function. Always returns 0.
-        """
-        return 0
-
-    def _reset_sim(self):
-        """ Reset to the initial sitting position.
-
-        Returns:
-            bool: `True`
-        """
-        # set qpos as new initial position and velocity as zero
-        qpos = self.init_qpos
-        qvel = np.zeros(self.sim.data.qvel.shape)
-
-        new_state = mujoco_py.MjSimState(
-            self.initial_state.time, qpos, qvel, self.initial_state.act, self.initial_state.udd_state
-        )
-
-        self.sim.set_state(new_state)
-        self.sim.forward()
-
-        return True
-
-    def _is_failure(self, achieved_goal, desired_goal):
-        """ Dummy function that always returns False.
-
-        Args:
-            achieved_goal (object): This parameter is ignored.
-            desired_goal (object): This parameter is ignored.
-
-        Returns:
-            bool: `False`
-        """
-        return False
-
-    def _get_achieved_goal(self):
-        """ Dummy function that returns an empty array.
-
-        Returns:
-            numpy.ndarray: An empty array.
-        """
-        return np.zeros(self.goal.shape)
-
-    def _viewer_setup(self):
-        """Initial configuration of the viewer. Can be used to set the camera position,
-        for example.
-        """
-        #self.viewer.cam.trackbodyid = 0  # id of the body to track
-        self.viewer.cam.distance = 1.5  # how much you "zoom in", smaller is closer
-        self.viewer.cam.lookat[0] = 0  # x,y,z offset from the object (works if trackbodyid=-1)
-        self.viewer.cam.lookat[1] = -0.3
-        self.viewer.cam.lookat[2] = 0.5  # 0.24 -0.04 .8
-        self.viewer.cam.elevation = 0
-        self.viewer.cam.azimuth = 180
+                         actuation_model=MuscleModel)
