+++ mimoProprioception/proprio.py
@@ -56,9 +56,11 @@ class SimpleProprioception(Proprioception):
     This class can provide relative joint positions, joint velocities, joint torques and limit sensors on the joint
     range of motion. Torques are in newton-meters, all other values in radians. Joint positions are always part of the
     output, while the others can be added optionally through the configuration dictionary. Valid components are
-    'velocity', 'torque', 'limits'.
+    'velocity', 'torque', 'limits' and 'actuation'.
     The limit sensing increases linearly from 0 through 1 and beyond as the joint position moves within the threshold
     distance to the limit and then exceeds the limit. The threshold is part of the configuration.
+    The 'actuation' component returns quantities from the actuation model. What these are depends on the specific
+    implementation.
     The configuration dictionary should have the form::
 
         {
@@ -67,7 +69,7 @@ class SimpleProprioception(Proprioception):
         }
 
     Joint positions, velocities and limits are read from the simulation state directly, joint torques uses torque
-    sensors placed between bodies in the scene. By default MIMo has one sensor for each joint. Any torque sensor with
+    sensors placed between bodies in the scene. By default, MIMo has one sensor for each joint. Any torque sensor with
     the 'proprio' prefix is used for the output.
 
     Attributes:
@@ -85,7 +87,7 @@ class SimpleProprioception(Proprioception):
     """
 
     #: Valid entries for the output components
-    VALID_COMPONENTS = ["velocity", "torque", "limits"]
+    VALID_COMPONENTS = ["velocity", "torque", "limits", "actuation"]
 
     def __init__(self, env, proprio_parameters):
         super().__init__(env, proprio_parameters)
@@ -114,7 +116,7 @@ class SimpleProprioception(Proprioception):
         if "limits" in self.output_components:
             self.sensor_names["limit"] = self.joint_names
 
-        if "threshold" in proprio_parameters:
+        if proprio_parameters is not None and "threshold" in proprio_parameters:
             self.limit_thresh = proprio_parameters["threshold"]
         else:
             self.limit_thresh = .035  # ~2 degrees in radians
@@ -149,5 +151,8 @@ class SimpleProprioception(Proprioception):
             response = np.minimum(l_dif, u_dif) / self.limit_thresh
             response = - np.minimum(response, 0)
             self.sensor_outputs["limits"] = response
+
+        if "actuation" in self.output_components:
+            self.sensor_outputs["actuation"] = self.env.actuation_model.observations().flatten()
 
         return np.concatenate([self.sensor_outputs[key] for key in sorted(self.sensor_outputs.keys())])
